Remove two warnings about incompatible pointer type
Implementation should conform to SqDisplay.h

platforms/unix/vm-display-X11/sqUnixX11.c:7453:1: warning: incompatible pointer types initializing 'long (*)(char *, int)' with an expression of type 'long (char *, long)' [-Wincompatible-pointer-types]SqDisplayDefine(X11);^~~~~~~~~~~~~~~~~~~~platforms/unix/vm/SqDisplay.h:108:3: note: expanded from macro 'SqDisplayDefine'  display_winImageFind,                         \  ^~~~~~~~~~~~~~~~~~~~/platforms/unix/vm-display-X11/sqUnixX11.c:7453:1: warning: incompatible pointer types initializing 'long (*)(unsigned int *, long, long, long, long, long, long, long, long)' with an expression of type 'long (unsigned int *, long, long, long, int, int, int, int, int)' [-Wincompatible-pointer-types]SqDisplayDefine(X11);^~~~~~~~~~~~~~~~~~~~platforms/unix/vm/SqDisplay.h:157:3: note: expanded from macro 'SqDisplayDefine'  display_hostWindowShowDisplay,                \  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ platforms/unix/vm-display-X11/sqUnixX11.c
@@ -7205,7 +7205,7 @@ static long display_hostWindowCreate(long w, long h, long x, long y, char *list,
 static long display_hostWindowClose(long index)                                               { return 0; }
 static long display_hostWindowCloseAll(void)                                                 { return 0; }
 static long display_hostWindowShowDisplay(unsigned *dispBitsIndex, long width, long height, long depth,
-					 int affectedL, int affectedR, int affectedT, int affectedB, int windowIndex)
+					 long affectedL, long affectedR, long affectedT, long affectedB, long windowIndex)
 											    { return 0; }
 
 /* By convention for HostWindowPlugin, handle 1 refers to the display window */
@@ -7369,7 +7369,7 @@ static void display_winExit(void)
 }
 
 
-static long  display_winImageFind(char *buf, long len)	{ return 0; }
+static long  display_winImageFind(char *buf, int len)	{ return 0; }
 static void display_winImageNotFound(void)		{}
 
 #if SqDisplayVersionMajor >= 1 && SqDisplayVersionMinor >= 3
